backend/requirements.txt
fastapi
uvicorn 
gitpython


openai
faiss-cpu
tiktoken
dotenv